Converts stealable artifacts to a system instead of an item
